+++ src/auth/login.ts
@@ -4,7 +4,7 @@ import {
   jsonContentRequired,
   jsonMessageContent,
 } from '@/lib/helper'
-import { generateAuthTokenAndSetCookie } from '@/utils/authToken'
+import { generateAuthTokenAndSetCookie } from '@/utils/auth-token'
 import { verifyPassword } from '@/utils/hash'
 import { createRoute, z } from '@hono/zod-openapi'
 import type { Context } from 'hono'
