+++ src/index.ts
@@ -2,7 +2,7 @@ import { loginHandler, loginRoute } from '@/auth/login'
 import { registerHandler, registerRoute } from '@/auth/register'
 import { authStatusHandler, authStatusRoute } from '@/auth/status'
 import { errorHook } from '@/lib/helper'
-import { authMiddleware } from '@/middleware/authMiddleware'
+import { authMiddleware } from '@/middleware/auth-middleware'
 import { notFound } from '@/middleware/not-found'
 import { onError } from '@/middleware/on-error'
 import { OpenAPIHono } from '@hono/zod-openapi'
