+++ src/lib/helper/json-content-required.ts
@@ -1,4 +1,4 @@
-import { jsonContent } from './jsonContent'
+import { jsonContent } from './json-content'
 import type { ZodSchema } from './types'
 
 export const jsonContentRequired = <T extends ZodSchema>(
