+++ src/lib/helper/json-message-content.ts
@@ -1,5 +1,5 @@
 import { z } from '@hono/zod-openapi'
-import { jsonContent } from './jsonContent'
+import { jsonContent } from './json-content'
 
 export const jsonMessageContent = (message: string) => {
   const messageSchema = z.object({
