+++ src/utils/password-validator.ts
